Include event source in global events path
Event IDs alone do not need to be globally unique. That is why the
source field exists in the first place. It is the combination of source
plus ID that must be unique. Therefore, there must be an additional
directory here to ensure events do not collide.

see: https://github.com/cloudevents/spec/blob/v1.0.2/cloudevents/spec.md#source-1